restructured Query.js
