+++ resolvers/direct_chat_query/directChatMessages.js
@@ -1,0 +1,5 @@
+const DirectChatMessages = require('../../models/DirectChatMessage');
+
+module.exports = async () => {
+  return await DirectChatMessages.find();
+};
