+++ resolvers/direct_chat_query/directChats.js
@@ -1,0 +1,5 @@
+const DirectChat = require('../../models/DirectChat');
+
+module.exports = async () => {
+  return await DirectChat.find();
+};
