+++ resolvers/group_chat_query/groupChatMessages.js
@@ -1,0 +1,5 @@
+const GroupChatMessages = require('../../models/GroupChatMessage');
+
+module.exports = async () => {
+  return await GroupChatMessages.find();
+};
