+++ resolvers/group_chat_query/groupChats.js
@@ -1,0 +1,5 @@
+const GroupChat = require('../../models/GroupChat');
+
+module.exports = async () => {
+  return await GroupChat.find();
+};
