+++ resolvers/group_query/groups.js
@@ -1,0 +1,5 @@
+const Group = require('../../models/Group');
+
+module.exports = async () => {
+  return await Group.find();
+};
